added handlers, 1 repository func, fixed some things

+++ internal/models/models.go
@@ -1,13 +1,13 @@
 package models
 
 type Discipline struct {
-	ID    int64  `db:"id" json:"id"`
+	ID    int    `db:"id" json:"id"`
 	Title string `db:"title" json:"title"`
 }
 
 type Document struct {
-	ID           int64  `db:"id" json:"id"`
-	DisciplineID int64  `db:"discipline_id" json:"discipline_id"`
+	ID           int    `db:"id" json:"id"`
+	DisciplineID int    `db:"discipline_id" json:"discipline_id"`
 	Title        string `db:"title" json:"title"`
 	Filepath     string `db:"file_path" json:"file_path"`
 }
